Completed stylings and responsive layout
