+++ src/App.css
@@ -1,16 +1,155 @@
 /* Styles for App.jsx */
 
+:root, #root {
+    font-family:'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+    background-color: #d7dce0;
+}
+
 h1, p {
     color: #24292f;
     margin: 10px;
 }
 
+/* Header component stylings */ 
+.headerName {
+    font-size: 3em;
+    text-decoration: none;
+}
 
+.nav-link {
+    color: #0a66c2;
+}
 
+/* About component stylings */ 
 .profilePicture {
-    width: 25%;
+    width: 50%;
     max-width: 300px;
+    height: 50%;
     border-radius: 30px;
     margin: 10px;
 }
 
+
+/* Project component stylings */ 
+.project {
+    border: 10px solid #24292f;
+    border-radius: 10px;
+}
+
+.projectOverlay {
+    display: flex;
+    justify-content: flex-end;
+    background-color: #d7dce0;
+    padding: 3px;
+    position: absolute;
+    bottom: 0%;
+    width: 100%;
+}
+
+.nameAnchor {
+    text-decoration: none;
+}
+
+.projectName {
+    color: #0a66c2;
+    text-wrap: nowrap;
+}
+
+.logoAnchor {
+    width: fit-content;
+    display: flex;
+    align-items: center;
+}
+
+.githubLogoProject {
+    width: 35px;
+    height: 35px; 
+}
+
+/* Contact component stylings */ 
+.contactForm {
+    margin: 10px;
+}
+
+.emailLink {
+    color: #0a66c2;
+    display: block;
+    margin-bottom: 20px;
+}
+
+input {
+    width: 325px;
+}
+
+.messageInput {
+    width: 325px;
+    height: 250px;
+}
+
+/* Resume component stylings */ 
+.resumeLink {
+    color: #0a66c2;
+}
+
+.skills {
+    margin: 10px;
+}
+
+/* Footer component stylings */ 
+.footerContainer {
+    margin-top: 5px;
+    display: flex;
+    justify-content: center;
+}
+
+.logoFooter {
+    max-width: 50px;
+    margin: 10px;
+}
+
+
+@media screen and (min-width: 780px) {
+    .headerName:hover {
+        color: #0a66c2;
+    }
+
+    .headerName:active {
+        color: #9bc5ee;
+    }
+
+    .aboutContent {
+        display: flex;
+    }
+
+    .projectGrid {
+        width: 100%;
+        display: flex;
+        flex-wrap: wrap;
+        margin-left: 1%;
+        margin-right: 1%;
+    }
+
+    .project {
+        width: 47%;
+    }
+
+    .project:hover {
+        border: 10px solid #d7dce0;
+    }
+
+      .project:active .projectName  {
+        color: #9bc5ee;
+    }
+
+    .emailLink:hover, .resumeLink:hover {
+        color: #9bc5ee;
+    }
+
+    .emailLink:active, .resumeLink:active  {
+        color: #24292f;
+    }
+   
+    .logo:hover {
+        opacity: 0.5;
+    }
+  }
