+++ src/App.jsx
@@ -1,8 +1,8 @@
 // Bringing in the required import from 'react-router-dom'
 import { Outlet } from 'react-router-dom';
 
-import Header from './components/Header/Header.jsx';
-import Footer from './components/Footer/Footer.jsx';
+import Header from './components/Header.jsx';
+import Footer from './components/Footer.jsx';
 import './App.css'
 
 function App() {
