+++ src/components/Header.jsx
@@ -1,11 +1,10 @@
 import { Link } from 'react-router-dom';
 
-import Navigation from '../Navigation/Navigation.jsx';
-import './Header.css';
+import Navigation from './Navigation.jsx';
 
 function Header() {
     return (
-        <div>
+        <div className="headerBackground">
             <div className="m-3">
             <Link className="h1 headerName" to="/">
             Brian Wach
