--- src/components/Header/Header.css
@@ -1,8 +1,0 @@
-.headerName {
-    font-size: 3em;
-    text-decoration: none;
-}
-
-.headerName:hover {
-    color: #0a66c2;
-}
