+++ src/components/Navigation.jsx
@@ -1,5 +1,4 @@
 import { Link, useLocation } from 'react-router-dom';
-import './Navigation.css';
 
 function Navigation() {
     const currentPage = useLocation().pathname;
