--- src/components/Navigation/Navigation.css
@@ -1,3 +1,0 @@
-.nav-item {
-
-}
