--- src/components/Project/Project.css
@@ -1,23 +1,0 @@
-.projectOverlay {
-    display: flex;
-    justify-content: flex-end;
-    background-color: #d7dce0;
-    padding: 10px;
-    position: absolute;
-    bottom: 0%;
-    width: 100%;
-}
-
-.projectName {
-    display: inline;
-    text-wrap: nowrap;
-}
-
-.logoAnchor {
-    width: fit-content;
-}
-
-.githubLogoProject {
-    max-width: 10%;
-    height: auto; 
-}
