--- src/index.css
@@ -1,9 +1,0 @@
-/* Styles for main.jsx */
-
-:root {
-    background-color: #d7dce0;
-}
-
-#root {
-    background-color: #d7dce0;
-}
