+++ src/pages/About.jsx
@@ -1,15 +1,17 @@
 export default function About() {
-    return (
-      <div>
-        <h1>About Me</h1>
-            <img className="profilePicture" src='/images/headshot.jpg' alt='Headshot of Brian Wach' />
+  return (
+    <div>
+      <h1>About Me</h1>
+      <div className="aboutContent">
+        <img className="profilePicture" src='/images/headshot.jpg' alt='Headshot of Brian Wach' />
         <p>
-        Brian Wach is an audio-visual technician, musician, and composer in the Chicagoland area. 
-        He completed his master’s degree in Recording Arts and earned a bachelor’s in Percussion Performance, both at Northern Illinois University (NIU). 
-        Brian strives to combine his passion for technology and music in order to create inspiring and meaningful experiences for others. 
-        Since 2014, he has been working in event production and information technologies as a graduate assistant at NIU, audio warehouse manager with Sound Investment, and freelance streaming technician/virtual meeting producer for Brella Productions and Intempio. 
-        Before the current health pandemic, Brian could also be found playing/recording music with his acoustic duo, Off the Curb, sound engineering for the Emanation Music Collective, and performing with the improv comedy team, All Caps.
+          Brian Wach is an audio-visual technician, musician, and composer in the Chicagoland area.
+          He completed his master’s degree in Recording Arts and earned a bachelor’s in Percussion Performance, both at Northern Illinois University (NIU).
+          Brian strives to combine his passion for technology and music in order to create inspiring and meaningful experiences for others.
+          Since 2014, he has been working in event production and information technologies as a graduate assistant at NIU, audio warehouse manager with Sound Investment, and freelance streaming technician/virtual meeting producer for Brella Productions and Intempio.
+          Before the current health pandemic, Brian could also be found playing/recording music with his acoustic duo, Off the Curb, sound engineering for the Emanation Music Collective, and performing with the improv comedy team, All Caps.
         </p>
       </div>
-    );
-  }
+    </div>
+  );
+}
